+++ tests/testthat/test-emissions_profile_any_at_product_level.R
@@ -134,13 +134,10 @@ test_that("if 'isic' column is numeric it knows how to handle it gracefully", {
   expect_no_error(emissions_profile_any_at_product_level(companies, products))
 })
 
-test_that("if the 'isic' column hasn't 4 digits throws an errors ", {
+test_that("if the 'isic' column hasn't 4 digits throws no error", {
   companies <- example_companies()
-  products <- example_products()
+  products <- example_products(!!aka("isic") := c(NA, "1", "12", "123", "12345"))
   expect_no_error(emissions_profile_any_at_product_level(companies, products))
-
-  products$isic_4digit <- "1"
-  expect_error(emissions_profile_any_at_product_level(companies, products), "must.*4")
 })
 
 test_that("a 0-row `co2` yields an error", {