+++ tests/testthat/test-jitter_range.R
@@ -34,3 +34,23 @@ test_that("is sensitive to `amount`", {
   expect_true(mean(out2$min_jitter) < mean(out1$min_jitter))
   expect_true(mean(out2$max_jitter) > mean(out1$max_jitter))
 })
+
+test_that("percent noise is even not monotonically decreasing", {
+  local_seed(123)
+  data <- tibble(min = c(0.1, 1, 10, 100, 1000), max = min)
+
+  out <- jitter_range(data, factor = 1, amount = NULL)
+
+  is_decreasing <- function(x) all(x == cummin(x))
+  expect_false(is_decreasing(percent_noise(out$min, out$min_jitter)))
+})
+
+test_that("jittering 0 yields non-0", {
+  local_seed(123)
+  data <- tibble(min = 0, max = min)
+
+  out <- jitter_range(data, factor = 1, amount = NULL)
+
+  expect_true(out$min_jitter != 0)
+  expect_true(out$max_jitter != 0)
+})